Final changes
Tidy stuff up and regret writing it in such a confusing way

+++ RomanCalc/src/romanNumerals.cpp
@@ -19,17 +19,17 @@ std::string int_to_roman(int integer) {
 	int roman_nums[7] = {0};
 
 	//Find out how many of each character we need by finding DIV
-	roman_nums[6] = integer / 1000;
+	roman_nums[6] = static_cast<int>(integer / 1000);
 	integer = integer - (roman_nums[6] * 1000);
-	roman_nums[5] = integer / 500;
+	roman_nums[5] = static_cast<int>(integer / 500);
 	integer = integer - (roman_nums[5] * 500);
-	roman_nums[4] = integer / 100;
+	roman_nums[4] = static_cast<int>(integer / 100);
 	integer = integer - (roman_nums[4] * 100);
-	roman_nums[3] = integer / 50;
+	roman_nums[3] = static_cast<int>(integer / 50);
 	integer = integer - (roman_nums[3] * 50);
-	roman_nums[2] = integer / 10;
+	roman_nums[2] = static_cast<int>(integer / 10);
 	integer = integer - (roman_nums[2] * 10);
-	roman_nums[1] = integer / 5;
+	roman_nums[1] = static_cast<int>(integer / 5);
 	roman_nums[0] = integer - (roman_nums[1] * 5);
 
 	//Create string output in correct order
@@ -44,42 +44,38 @@ std::string int_to_roman(int integer) {
 }
 
 int roman_to_int(const std::string& roman) {
+	//Function to turn a roman numeral string to an integer
 	int roman_len = roman.length();
 	int value = 0;
+
 	for (int i = 0; i < roman_len; i++) {
+
 		switch (roman[i]) {
-		case 'I':
-		case 'i':
+		case 'I': case 'i':
 			value = value + 1;
 			break;
 
-		case 'V':
-		case 'v':
+		case 'V': case 'v':
 			value = value + 5;
 			break;
 
-		case 'X':
-		case 'x':
+		case 'X': case 'x':
 			value = value + 10;
 			break;
 
-		case 'L':
-		case 'l':
+		case 'L': case 'l':
 			value = value + 50;
 			break;
 
-		case 'C':
-		case 'c':
+		case 'C': case 'c':
 			value = value + 100;
 			break;
 
-		case 'D':
-		case 'd':
+		case 'D': case 'd':
 			value = value + 500;
 			break;
 
-		case 'M':
-		case 'm':
+		case 'M': case 'm':
 			value = value + 1000;
 			break;
 
@@ -87,5 +83,6 @@ int roman_to_int(const std::string& roman) {
 			break;
 		}
 	}
+
 	return value;
 }
